[MOV] tests: move gauge and scorecard tests to their own folders
Task: 4170179
